+++ core/src/main/java/be/orbinson/aem/dictionarytranslator/services/impl/DictionaryServiceImpl.java
@@ -1,45 +1,91 @@
 package be.orbinson.aem.dictionarytranslator.services.impl;
 
-import be.orbinson.aem.dictionarytranslator.exception.DictionaryException;
-import be.orbinson.aem.dictionarytranslator.services.DictionaryService;
-import be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants;
-import com.day.cq.commons.jcr.JcrConstants;
-import com.day.cq.commons.jcr.JcrUtil;
-import com.day.cq.replication.ReplicationActionType;
-import com.day.cq.replication.ReplicationException;
-import com.day.cq.replication.Replicator;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.jackrabbit.util.Text;
-import org.apache.sling.api.resource.*;
-import org.apache.sling.jcr.resource.api.JcrResourceConstants;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.SLING_KEY;
+import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.SLING_MESSAGE;
+import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.SLING_MESSAGEENTRY;
+import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.SLING_MESSAGE_MIXIN;
+import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.MIX_LANGUAGE;
+import static org.apache.jackrabbit.JcrConstants.JCR_CONTENT;
+import static org.apache.jackrabbit.JcrConstants.JCR_LANGUAGE;
+import static org.apache.jackrabbit.JcrConstants.JCR_MIXINTYPES;
+import static org.apache.jackrabbit.JcrConstants.JCR_PRIMARYTYPE;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.TreeMap;
+import java.util.TreeSet;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.stream.Collectors;
 
 import javax.jcr.RepositoryException;
 import javax.jcr.Session;
 import javax.jcr.security.AccessControlManager;
 import javax.jcr.security.Privilege;
-import java.util.*;
-import java.util.concurrent.atomic.AtomicReference;
 
-import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.*;
-import static org.apache.jackrabbit.JcrConstants.*;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.jackrabbit.commons.json.JsonHandler;
+import org.apache.jackrabbit.commons.json.JsonParser;
+import org.apache.jackrabbit.util.Text;
+import org.apache.sling.api.resource.ModifiableValueMap;
+import org.apache.sling.api.resource.PersistenceException;
+import org.apache.sling.api.resource.Resource;
+import org.apache.sling.api.resource.ResourceMetadata;
+import org.apache.sling.api.resource.ResourceResolver;
+import org.apache.sling.api.resource.ResourceUtil;
+import org.apache.sling.api.resource.ValueMap;
+import org.apache.sling.api.resource.observation.ResourceChange;
+import org.apache.sling.api.resource.observation.ResourceChangeListener;
+import org.apache.sling.jcr.resource.api.JcrResourceConstants;
+import org.jetbrains.annotations.NotNull;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-@Component
-public class DictionaryServiceImpl implements DictionaryService {
+import com.day.cq.commons.jcr.JcrConstants;
+import com.day.cq.commons.jcr.JcrUtil;
+import com.day.cq.replication.ReplicationActionType;
+import com.day.cq.replication.ReplicationException;
+import com.day.cq.replication.Replicator;
+
+import be.orbinson.aem.dictionarytranslator.exception.DictionaryException;
+import be.orbinson.aem.dictionarytranslator.services.DictionaryService;
+import be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants;
+
+@Component(property = { ResourceChangeListener.PATHS + "=/",
+ResourceChangeListener.CHANGES + "=ADDED" ,
+ResourceChangeListener.CHANGES + "=REMOVED",
+ResourceChangeListener.CHANGES + "=CHANGED"})
+public class DictionaryServiceImpl implements DictionaryService, ResourceChangeListener {
 
     private static final Logger LOG = LoggerFactory.getLogger(DictionaryServiceImpl.class);
     private static final String SLING_BASENAME = "sling:basename";
+    // same as https://github.com/apache/sling-org-apache-sling-i18n/blob/3f98ebf430e416226500c2975086423edc29dcb3/src/main/java/org/apache/sling/i18n/impl/JcrResourceBundle.java#L69
+    static final String QUERY_LANGUAGE_ROOTS = "//element(*,mix:language)[@jcr:language]";
 
     @Reference
     private Replicator replicator;
 
+    /**
+     * Cache of language dictionaries, the key is an artifical path in the format {@code <dictionaryPath>/<language>}.
+     * The cache for all languages is automatically invalidated when a resource below the dictionary path was changed.
+     */
+    private final Map<String, Map<String, Message>> messagesPerLanguageDictionary = new HashMap<>();
+
     public boolean isEditableDictionary(Resource resource) {
         String path = resource.getPath();
+        if (getType(resource) != DictionaryType.SLING_MESSAGE_ENTRY) {
+            return false;
+        }
         Session session = resource.getResourceResolver().adaptTo(Session.class);
         if (session != null) {
             try {
@@ -66,8 +112,8 @@ public class DictionaryServiceImpl implements DictionaryService {
         ResourceResolver resourceResolver = dictionary.getResourceResolver();
 
         properties.put(ResourceResolver.PROPERTY_RESOURCE_TYPE, JcrResourceConstants.NT_SLING_FOLDER);
-        properties.put("jcr:language", language);
-        properties.put("jcr:mixinTypes", "mix:language");
+        properties.put(JCR_LANGUAGE, language);
+        properties.put(JCR_MIXINTYPES, MIX_LANGUAGE);
 
         if (StringUtils.isNotEmpty(basename)) {
             properties.put(SLING_BASENAME, basename);
@@ -84,7 +130,7 @@ public class DictionaryServiceImpl implements DictionaryService {
         Map<String, Resource> result = new TreeMap<>();
 
         resourceResolver
-                .findResources("//element(*, mix:language)[@jcr:language and (@jcr:primaryType='sling:Folder' or @jcr:primaryType='nt:folder')]/..", "xpath")
+                .findResources(QUERY_LANGUAGE_ROOTS + "/..", "xpath")
                 .forEachRemaining(resource -> result.put(resource.getPath(), resource));
 
         return new ArrayList<>(result.values());
@@ -94,7 +140,7 @@ public class DictionaryServiceImpl implements DictionaryService {
         LOG.debug("Create dictionary '{}'", name);
         ResourceResolver resourceResolver = parent.getResourceResolver();
         String dictionaryPath = String.format("%s/%s/i18n", parent.getPath(), JcrUtil.createValidName(name));
-        Resource dictionaryResource = ResourceUtil.getOrCreateResource(resourceResolver, dictionaryPath, "sling:Folder", "sling:Folder", true);
+        Resource dictionaryResource = ResourceUtil.getOrCreateResource(resourceResolver, dictionaryPath, JcrResourceConstants.NT_SLING_FOLDER, JcrResourceConstants.NT_SLING_FOLDER, true);
 
         for (String language : languages) {
             addLanguage(dictionaryResource, language, basename);
@@ -102,29 +148,25 @@ public class DictionaryServiceImpl implements DictionaryService {
     }
 
     @Override
-    public void deleteDictionary(ResourceResolver resourceResolver, String dictionaryPath) throws DictionaryException {
+    public void deleteDictionary(ResourceResolver resourceResolver, String dictionaryPath) throws DictionaryException, ReplicationException, PersistenceException {
         LOG.debug("Delete dictionary '{}'", dictionaryPath);
-        try {
-            final Resource dictionaryResource = resourceResolver.getResource(dictionaryPath);
-            if (dictionaryResource != null) {
-                replicator.replicate(resourceResolver.adaptTo(Session.class), ReplicationActionType.DEACTIVATE, dictionaryResource.getPath());
-                resourceResolver.delete(dictionaryResource);
-                resourceResolver.commit();
-            } else {
-                throw new DictionaryException("Dictionary '" + dictionaryPath + "' not found");
-            }
-        } catch (PersistenceException | ReplicationException e) {
-            throw new DictionaryException("Could not delete dictionary", e);
+        final Resource dictionaryResource = resourceResolver.getResource(dictionaryPath);
+        if (dictionaryResource != null) {
+            replicator.replicate(resourceResolver.adaptTo(Session.class), ReplicationActionType.DEACTIVATE, dictionaryResource.getPath());
+            resourceResolver.delete(dictionaryResource);
+            resourceResolver.commit();
+        } else {
+            throw new DictionaryException("Dictionary '" + dictionaryPath + "' not found");
         }
     }
-
 
     public List<String> getLanguages(Resource dictionaryResource) {
         Set<String> result = new TreeSet<>();
 
         dictionaryResource.listChildren().forEachRemaining(child -> {
             ValueMap properties = child.getValueMap();
-            if (properties.containsKey(JCR_LANGUAGE)) {
+            // mixin check not implemented due to https://issues.apache.org/jira/browse/SLING-12779
+            if (properties.containsKey(JCR_LANGUAGE) /* && hasMixinType(child, MIX_LANGUAGE)*/) {
                 LOG.trace("Found language with path '{}'", child.getPath());
                 result.add(properties.get(JCR_LANGUAGE, String.class));
             }
@@ -134,20 +176,13 @@ public class DictionaryServiceImpl implements DictionaryService {
     }
 
     @Override
-    public void deleteLanguage(ResourceResolver resourceResolver, Resource dictionaryResource, String language) throws DictionaryException {
-        Resource languageResource = getLanguageResource(dictionaryResource, language);
-        if (languageResource != null) {
-            try {
-                LOG.debug("Delete language '{}' from '{}'", language, dictionaryResource.getPath());
-                replicator.replicate(resourceResolver.adaptTo(Session.class), ReplicationActionType.DEACTIVATE, languageResource.getPath());
-                resourceResolver.delete(languageResource);
-                resourceResolver.commit();
-            } catch (PersistenceException | ReplicationException e) {
-                throw new DictionaryException("Could not delete language", e);
-            }
-        } else {
-            throw new DictionaryException("Language does not exist: " + language);
-        }
+    public void deleteLanguage(Resource dictionaryResource, String language) throws DictionaryException, ReplicationException, PersistenceException {
+        Resource languageResource = getLanguageResource(dictionaryResource, language).orElseThrow(() -> new DictionaryException("Language resource not found for language \"" + language + "\" of dictionary \"" + dictionaryResource.getPath() + "\""));
+        LOG.debug("Delete language '{}' from '{}'", language, dictionaryResource.getPath());
+        ResourceResolver resourceResolver = dictionaryResource.getResourceResolver();
+        replicator.replicate(resourceResolver.adaptTo(Session.class), ReplicationActionType.DEACTIVATE, languageResource.getPath());
+        resourceResolver.delete(languageResource);
+        resourceResolver.commit();
     }
 
     @Override
@@ -170,74 +205,39 @@ public class DictionaryServiceImpl implements DictionaryService {
      * @param language           The language
      * @return the language resource if it exists
      */
-    @Override
-    public @Nullable Resource getLanguageResource(Resource dictionaryResource, String language) {
-        if (dictionaryResource != null) {
-            for (Resource languageResource : dictionaryResource.getChildren()) {
-                if (language.equals(languageResource.getValueMap().get(JcrConstants.JCR_LANGUAGE))) {
-                    return languageResource;
-                }
+    private Optional<Resource> getLanguageResource(Resource dictionaryResource, String language) {
+        Objects.requireNonNull(dictionaryResource, "dictionaryResource must not be null");
+        for (Resource languageResource : dictionaryResource.getChildren()) {
+            if (language.equals(languageResource.getValueMap().get(JcrConstants.JCR_LANGUAGE))) {
+                return Optional.of(languageResource);
             }
         }
-        return null;
+        return Optional.empty();
     }
 
     @Override
-    public List<String> getKeys(Resource dictionaryResource) {
+    public List<String> getKeys(Resource dictionaryResource) throws DictionaryException {
         Set<String> keys = new TreeSet<>();
+        // collect keys from all languages
         for (String language : getLanguages(dictionaryResource)) {
-            Resource languageResource = getLanguageResource(dictionaryResource, language);
-            if (languageResource != null) {
-                for (Resource messageEntryResource : languageResource.getChildren()) {
-                    if (isMessageEntryResource(messageEntryResource)) {
-                        String key = Optional.ofNullable(messageEntryResource.getValueMap().get(SLING_KEY, String.class))
-                                .orElse(messageEntryResource.getName());
-                        keys.add(key);
-                    }
-                }
-            }
+            keys.addAll(getMessages(dictionaryResource, language).keySet());
         }
         return List.copyOf(keys);
     }
 
-
     @Override
-    public boolean keyExists(Resource dictionaryResource, String language, String key) {
-        Resource languageResource = getLanguageResource(dictionaryResource, language);
-        return languageResource != null && getMessageEntryResource(languageResource, key) != null;
+    public boolean keyExists(Resource dictionaryResource, String language, String key) throws DictionaryException {
+        return getMessages(dictionaryResource, language).containsKey(key);
     }
 
     @Override
-    public void createMessageEntry(ResourceResolver resourceResolver, Resource dictionaryResource, String language, String key, String message) throws PersistenceException {
-        Resource languageResource = getLanguageResource(dictionaryResource, language);
-
-        if (languageResource != null) {
-            String path = languageResource.getPath();
-            Map<String, Object> properties = new HashMap<>();
-            properties.put(JCR_PRIMARYTYPE, SLING_MESSAGEENTRY);
-            properties.put(SLING_KEY, key);
-            if (!message.isBlank()) {
-                properties.put(SLING_MESSAGE, message);
-            }
-            resourceResolver.create(languageResource, Text.escapeIllegalJcrChars(key), properties);
-            resourceResolver.commit();
-            LOG.trace("Created message entry with key '{}' and message '{}' on path '{}'", key, message, path);
-        }
+    public void createOrUpdateMessageEntry(Resource dictionaryResource, String language, String key, String message) throws PersistenceException, DictionaryException {
+        Resource languageResource = getLanguageResource(dictionaryResource, language).orElseThrow(() -> new DictionaryException("Language resource not found for language \"" + language + "\" of dictionary \"" + dictionaryResource.getPath() + "\""));
+        Resource messageEntryResource = getOrCreateMessageEntryResource(languageResource, key);
+        updateMessage(key, message, messageEntryResource);
     }
 
-    @Override
-    public void updateMessageEntry(ResourceResolver resourceResolver, Resource dictionaryResource, String language, String key, String message) throws PersistenceException {
-        Resource languageResource = getLanguageResource(dictionaryResource, language);
-        if (languageResource != null) {
-            Resource messageEntryResource = getOrCreateMessageEntryResource(resourceResolver, languageResource, key);
-            if (messageEntryResource != null) {
-                updateMessage(key, message, messageEntryResource);
-            }
-            resourceResolver.commit();
-        }
-    }
-
-    private static void updateMessage(String key, String message, Resource messageEntryResource) {
+    private static void updateMessage(String key, String message, Resource messageEntryResource) throws PersistenceException {
         ValueMap valueMap = messageEntryResource.adaptTo(ModifiableValueMap.class);
         if (valueMap != null) {
             if (message.isBlank()) {
@@ -249,55 +249,206 @@ public class DictionaryServiceImpl implements DictionaryService {
                 }
                 LOG.trace("Updated message entry with name '{}' and message '{}' on path '{}'", messageEntryResource.getName(), message, messageEntryResource.getPath());
             }
+        } else {
+            throw new PersistenceException("Could not update message entry, resource at \"" + messageEntryResource.getPath() + "\" not adaptable to ModifiableValueMap");
         }
     }
 
-    @Override
-    public Resource getMessageEntryResource(Resource languageResource, String key) {
+    private Optional<Resource> getMessageEntryResource(Resource languageResource, String key) {
+        Objects.requireNonNull(languageResource, "languageResource must not be null");
+        Objects.requireNonNull(key, "key must not be null");
         // In order to speed up the search, we go for the default check where it is the escaped key as node name
         Resource messageEntryResource = languageResource.getChild(Text.escapeIllegalJcrChars(key));
         if (isMessageEntryResource(messageEntryResource)) {
-            return messageEntryResource;
+            return Optional.of(messageEntryResource);
         }
 
         // Fall back to searching for the resource with sling:key as correct property
         for (Resource resource : languageResource.getChildren()) {
             if (key.equals(resource.getValueMap().get(DictionaryConstants.SLING_KEY)) && isMessageEntryResource(messageEntryResource)) {
-                return resource;
+                return Optional.of(resource);
             }
         }
 
-        return null;
-    }
-
-    private static boolean isMessageEntryResource(Resource messageEntryResource) {
-        return messageEntryResource != null && (messageEntryResource.isResourceType(DictionaryConstants.SLING_MESSAGEENTRY) ||
-                Arrays.asList(messageEntryResource.getValueMap().get(JCR_MIXINTYPES, new String[0])).contains(SLING_MESSAGE_MIXIN));
+        return Optional.empty();
     }
 
     @Override
-    public void deleteMessageEntry(ResourceResolver resourceResolver, Resource combiningMessageEntryResource) throws PersistenceException, ReplicationException {
-        ValueMap properties = combiningMessageEntryResource.getValueMap();
-        if (properties.containsKey(CombiningMessageEntryResourceProvider.MESSAGE_ENTRY_PATHS)) {
-            for (String messageEntryPath : properties.get(CombiningMessageEntryResourceProvider.MESSAGE_ENTRY_PATHS, new String[0])) {
-                Resource messageEntryResource = resourceResolver.getResource(messageEntryPath);
-                if (messageEntryResource != null) {
-                    replicator.replicate(resourceResolver.adaptTo(Session.class), ReplicationActionType.DEACTIVATE, messageEntryPath);
-                    resourceResolver.delete(messageEntryResource);
-                }
-            }
-            resourceResolver.commit();
+    public DictionaryType getType(Resource dictionaryResource, String language) throws DictionaryException {
+        Resource languageResource = getLanguageResource(dictionaryResource, language).orElseThrow(() -> new DictionaryException("Language resource not found for language \"" + language + "\" in dictionary \"" + dictionaryResource.getPath() + "\""));
+        if (isJsonFileBasedDictionary(languageResource)) {
+            return DictionaryType.JSON_FILE;
+        } else {
+            return DictionaryType.SLING_MESSAGE_ENTRY;
         }
     }
 
-    private Resource getOrCreateMessageEntryResource(ResourceResolver resourceResolver, Resource languageResource, String key) throws PersistenceException {
-        Resource messageEntryResource = getMessageEntryResource(languageResource, key);
-        if (messageEntryResource != null) {
-            return messageEntryResource;
+    @Override
+    public DictionaryType getType(Resource dictionaryResource) {
+        DictionaryType type = null;
+        for (String language : getLanguages(dictionaryResource)) {
+            try {
+                DictionaryType newType = getType(dictionaryResource, language);
+                if (type != null && type != newType) {
+                    return DictionaryType.MIXED;
+                } else {
+                    type = newType;
+                }
+            } catch (DictionaryException e) {
+                LOG.warn("Could not get dictionary type for language '{}' below '{}, skipping language", language, dictionaryResource.getPath(), e);
+            }
         }
-        resourceResolver.create(languageResource, Text.escapeIllegalJcrChars(key), Map.of("jcr:primaryType", SLING_MESSAGEENTRY));
-        resourceResolver.commit();
-        return languageResource.getChild(Text.escapeIllegalJcrChars(key));
+        if (type == null) {
+            // for empty dictionaries, we return the default type
+            return DictionaryType.SLING_MESSAGE_ENTRY;
+        }
+        return type;
+    }
+
+    @Override
+    public void onChange(final @NotNull List<ResourceChange> changes) {
+        for (final ResourceChange change : changes) {
+            // always invalidate all languages of a dictionary
+            synchronized(this) {
+                messagesPerLanguageDictionary.keySet().removeIf(
+                    key -> {
+                        String dictionaryPath = Text.getRelativeParent(key, 1);
+                        if (change.getPath().startsWith(dictionaryPath)) {
+                            LOG.debug("Invalidating dictionary cache for path '{}'", dictionaryPath);
+                            return true;
+                        }
+                        return false;
+                    });
+            }
+        }
+    }
+ 
+    @Override
+    public synchronized Map<String, Message> getMessages(Resource dictionaryResource, String language) throws DictionaryException {
+        Map<String, Message> messages = messagesPerLanguageDictionary.get(dictionaryResource.getPath() + "/" + language);
+        if (messages == null) {
+            messages = loadMessages(dictionaryResource, language);
+            messagesPerLanguageDictionary.put(dictionaryResource.getPath() + "/" + language, messages);
+        }
+        return messages;
+    }
+
+    private Map<String, Message> loadMessages(Resource dictionaryResource, String language) throws DictionaryException {
+        Resource resource = getLanguageResource(dictionaryResource, language).orElseThrow(() -> new DictionaryException("Language resource not found for language \"" + language + "\" of dictionary \"" + dictionaryResource.getPath() + "\""));
+        if (isJsonFileBasedDictionary(resource)) {
+            Map<String, Object> messages = new HashMap<>();
+            loadJsonDictionary(resource.getChild(JCR_CONTENT), messages);
+            return messages.entrySet().stream()
+                    .collect(Collectors.toMap(Entry::getKey, e -> new Message(e.getValue().toString(), null)));
+        } else {
+            Map<String, Message> messages = new HashMap<>();
+            resource.listChildren().forEachRemaining(messageEntryResource -> {
+                if (isMessageEntryResource(messageEntryResource)) {
+                    String key = Optional.ofNullable(messageEntryResource.getValueMap().get(SLING_KEY, String.class))
+                            .orElse(Text.unescapeIllegalJcrChars(messageEntryResource.getName()));
+                    messages.put(key, new Message(messageEntryResource.getValueMap().get(SLING_MESSAGE, ""), messageEntryResource.getPath()));
+                }
+            });
+            return messages;
+        }
+    }
+
+    private boolean isJsonFileBasedDictionary(Resource languageResource) {
+        return (languageResource.getName().endsWith(".json") && languageResource.getChild(JCR_CONTENT) != null);
+    }
+
+    // start copy of https://github.com/apache/sling-org-apache-sling-i18n/blob/3f98ebf430e416226500c2975086423edc29dcb3/src/main/java/org/apache/sling/i18n/impl/JcrResourceBundle.java#L245
+    private void loadJsonDictionary(Resource resource, final Map<String, Object> targetDictionary) {
+        LOG.info("Loading json dictionary: {}", resource.getPath());
+
+        // use streaming parser (we don't need the dict in memory twice)
+        JsonParser parser = new JsonParser(new JsonHandler() {
+
+            private String key;
+
+            @Override
+            public void key(String key) throws IOException {
+                this.key = key;
+            }
+
+            @Override
+            public void value(String value) throws IOException {
+                targetDictionary.put(key, value);
+            }
+
+            @Override
+            public void object() throws IOException {}
+
+            @Override
+            public void endObject() throws IOException {}
+
+            @Override
+            public void array() throws IOException {}
+
+            @Override
+            public void endArray() throws IOException {}
+
+            @Override
+            public void value(boolean value) throws IOException {}
+
+            @Override
+            public void value(long value) throws IOException {}
+
+            @Override
+            public void value(double value) throws IOException {}
+        });
+
+        final InputStream stream = resource.adaptTo(InputStream.class);
+        if (stream != null) {
+            String encoding = "utf-8";
+            final ResourceMetadata metadata = resource.getResourceMetadata();
+            if (metadata.getCharacterEncoding() != null) {
+                encoding = metadata.getCharacterEncoding();
+            }
+
+            try {
+
+                parser.parse(stream, encoding);
+
+            } catch (IOException e) {
+                LOG.warn("Could not parse i18n json dictionary {}: {}", resource.getPath(), e.getMessage());
+            } finally {
+                try {
+                    stream.close();
+                } catch (IOException ignore) {
+                }
+            }
+        } else {
+            LOG.warn("Not a json file: {}", resource.getPath());
+        }
+    }
+
+    private static boolean isMessageEntryResource(Resource messageEntryResource) {
+        return messageEntryResource != null && (messageEntryResource.isResourceType(SLING_MESSAGEENTRY) ||
+                hasMixinType(messageEntryResource, SLING_MESSAGE_MIXIN));
+    }
+
+    private static boolean hasMixinType(Resource resource, String mixinType) {
+        return Arrays.asList(resource.getValueMap().get(JCR_MIXINTYPES, new String[0])).contains(mixinType);
+    }
+
+    @Override
+    public void deleteMessageEntry(Resource dictionaryResource, String language, String key) throws PersistenceException, ReplicationException, DictionaryException {
+        Resource languageResource = getLanguageResource(dictionaryResource, language).orElseThrow(() -> new DictionaryException("Language resource not found for language \"" + language + "\" of dictionary \"" + dictionaryResource.getPath() + "\""));
+        Resource messageEntryResource = getMessageEntryResource(languageResource, key).orElseThrow(() -> new DictionaryException("Message entry resource not found for key \"" + key + "\" in language \"" + language + "\" of dictionary \"" + dictionaryResource.getPath() + "\""));
+        replicator.replicate(dictionaryResource.getResourceResolver().adaptTo(Session.class), ReplicationActionType.DEACTIVATE, messageEntryResource.getPath());
+        dictionaryResource.getResourceResolver().delete(messageEntryResource);
+    }
+
+    private @NotNull Resource getOrCreateMessageEntryResource(Resource languageResource, String key) throws PersistenceException {
+        Optional<Resource> messageEntryResource = getMessageEntryResource(languageResource, key);
+        if (messageEntryResource.isPresent()) {
+            return messageEntryResource.get();
+        } else {
+            Resource newResource = languageResource.getResourceResolver().create(languageResource, Text.escapeIllegalJcrChars(key), Map.of(JCR_PRIMARYTYPE, SLING_MESSAGEENTRY));
+            LOG.trace("Created message entry with key '{}' on path '{}'", key, newResource.getPath());
+            return newResource;
+        }
     }
 
 }