+++ core/src/main/java/be/orbinson/aem/dictionarytranslator/servlets/action/UpdateMessageEntryServlet.java
@@ -46,8 +46,8 @@ public class UpdateMessageEntryServlet extends SlingAllMethodsServlet {
             htmlResponse.send(response, true);
         } else {
             ResourceResolver resourceResolver = request.getResourceResolver();
-            Resource combiningMessageEntryResource = resourceResolver.getResource(combiningMessageEntryPath);
             try {
+                Resource combiningMessageEntryResource = resourceResolver.getResource(combiningMessageEntryPath);
                 if (combiningMessageEntryResource != null) {
                     // javasecurity:S5145
                     LOG.debug("Update message entry for path '{}'", combiningMessageEntryPath);
@@ -77,8 +77,9 @@ public class UpdateMessageEntryServlet extends SlingAllMethodsServlet {
                 if (message == null) {
                     throw new DictionaryException("Unable to get message for language '" + language + "'");
                 }
-                dictionaryService.updateMessageEntry(resourceResolver, dictionaryResource, language, key, message);
+                dictionaryService.createOrUpdateMessageEntry(dictionaryResource, language, key, message);
             }
+            resourceResolver.commit();
         } else {
             throw new DictionaryException("Could not find dictionary path");
         }